refactor: services/file_analysis.py → file_analysis_service.py 로 rename (#260)
