+++ apps/ai/routers/file_analysis.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter
 
 from schemas.api.file_analysis import FileAnalysisRequest, FileAnalysisResponse
-from services.file_analysis import analyze_file
+from services.file_analysis_service import analyze_file
 
 router = APIRouter(prefix="/v1/llm", tags=["01 file analysis"])
